refactor(pdf-parser): add tests for PythonEnvironment and DoclingServer setup flow
Add comprehensive unit tests for `PythonEnvironment` and `DoclingServer` to validate setup steps, process handling, and error scenarios. Enhance reliability and coverage for environment and server initialization logic.
